refactor: remove Python 2 support, require Python 3.8+
- Removed all Python 2 compatibility code from py3compat.py
- Removed long() type conversions from number.py (Python 3 int handles large numbers)
- Added typer dependency for CLI functionality
- Updated setup.py and pyproject.toml with typer dependency
- Added console_scripts entry point for libcrypto CLI
- All 117 tests passing
- CLI functionality verified (generate, mnemonic commands working)

+++ src/libcrypto/cryptod/Crypto/Util/number.py
@@ -279,8 +279,8 @@ def getStrongPrime(N, e=0, false_positive_prob=1e-6, randfunc=None):
     # Use the accelerator if available
     if _fastmath is not None:
         return _fastmath.getStrongPrime(
-            long(N),  # noqa: F821 - long exists in Python 2
-            long(e),  # noqa: F821 - long exists in Python 2
+            N,
+            e,
             false_positive_prob,
             randfunc,
         )
@@ -400,7 +400,7 @@ def isPrime(N, false_positive_prob=1e-6, randfunc=None):
 
     if _fastmath is not None:
         return _fastmath.isPrime(
-            long(N),  # noqa: F821 - long exists in Python 2
+            N,
             false_positive_prob,
             randfunc,
         )
